requirements.txt
anyio==3.7.0
certifi==2023.5.7
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.96.0
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
ipython==8.21.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
orjson==3.9.0
pydantic==1.10.8
python-dotenv==1.0.0
python-multipart==0.0.6
python-telegram-bot==20.4
PyYAML==6.0
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
ujson==5.7.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
